feat: Support custom dialog in UIKit
